Finish

+++ Assets/Scripts/Data/BoxData.cs
@@ -6,12 +6,12 @@ public class BoxData
 {
     public int id;
     public string spriteName;
-    public float health;
-    public int coins;
+    public double health;
+    public long coins;
     public int experience;
     public float coolDown;
 
-    public BoxData(int id, string spriteName, float health, int coins, int experience, float coolDown)
+    public BoxData(int id, string spriteName, double health, long coins, int experience, float coolDown)
     {
         this.id = id;
         this.spriteName = spriteName;
